+++ fabric-item-api-v1/src/main/java/net/fabricmc/fabric/mixin/elytra/LivingEntityMixin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.fallflying;
+package net.fabricmc.fabric.mixin.elytra;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -26,17 +26,17 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 
-import net.fabricmc.fabric.api.item.v1.fallflying.FabricFallFlyingItem;
+import net.fabricmc.fabric.api.item.v1.FabricElytraFlyingItem;
 
 @Mixin(LivingEntity.class)
 public class LivingEntityMixin {
 	@Redirect(method = "initAi", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;getItem()Lnet/minecraft/item/Item;"))
 	private Item redirectGetItem(ItemStack stack) {
-		return stack.getItem() instanceof FabricFallFlyingItem ? Items.ELYTRA : stack.getItem();
+		return stack.getItem() instanceof FabricElytraFlyingItem ? Items.ELYTRA : stack.getItem();
 	}
 
 	@Redirect(method = "initAi", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ElytraItem;isUsable(Lnet/minecraft/item/ItemStack;)Z"))
 	private boolean redirectIsUsable(ItemStack stack) {
-		return stack.getItem() instanceof FabricFallFlyingItem ? ((FabricFallFlyingItem) stack.getItem()).shouldAllowFallFlying(stack, (LivingEntity) (Object) this) : ElytraItem.isUsable(stack);
+		return stack.getItem() instanceof FabricElytraFlyingItem ? ((FabricElytraFlyingItem) stack.getItem()).shouldAllowElytraFlight(stack, (LivingEntity) (Object) this) : ElytraItem.isUsable(stack);
 	}
 }
