+++ fabric-item-api-v1/src/main/resources/fabric-item-api-v1-elytra.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "net.fabricmc.fabric.mixin.fallflying",
+  "package": "net.fabricmc.fabric.mixin.elytra",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "LivingEntityMixin",
